Doom batch file generator
This is a little script that generates a sequence of batch files for
playing Doom with pistol starts on each level.

Each game or mod is defined in a little CSV file that the script reads
and it uses that info to generate a set of batch files based on the
particular game.