Search for source ports to choose
Searches for source ports in the DOOM_HOME\source-ports directory. If
the WINDOWS_DOOM_HOME and UNIX_DOOM_HOME variables are not set, it will
look for a single user under C:\Users\ and then assume that DOOM_HOME is
in C:\Users\<username>\doom.

The constructors of the other source ports will need to be modified for
this new method.