Generate batch files for Crispy Doom
This changes the source port to return multiple commands for the batch
file, because of the weird way Crispy Doom works with config files. Even
though it has a -config option, it seems it needs to be relative to the
Crispy Doom directory. Each batch file will copy the custom config file
into the game directory, then delete it at the end (just so that the
source port directory is left clean).

This also configures the batch file generation to use the launchers
directory and more information from the game to generate a batch file
with the following directory structure:
<DOOM_PATH>/launchers/<RELEASE_DATE> -- <GAME_NAME>/<PORT_NAME>/MAPXX -- EXXMXX -- <MISSION_NAME>.bat

For example:
/mnt/c/Users/Chris/doom/launchers/1993-12-10 -- DOOM/crispy/MAP01 -- E01M01 -- Hanger.bat

The game data files are updated to include the release date of the game,
and there's now also doom.csv and ultimate-doom.csv, since they were
actually 2 separate releases.