Recording configuration for GZDoom
Adds a 'record' configuration. If you run this, it will record a demo
with the name 'MAPX-YYYY-MM-DD-TIME' where time is the current hour and
minute. That file is saved in the same directory as the source port, but
a post game command will move it to the demos directory at DOOM_HOME.