Generate Batch for Starting with No Warp
Changes the launchers generation to also produce a 'start.bat' file that
will start the game without warping to a particular map. This is useful
for using OBS to record pre-made demos because you can launch the game
and then use the console to play the demo. I couldn't get OBS to stop
producing a blank screen for 10 seconds when it initially launched the
game, so having the manual control of initiating the demo means you can
record it properly.