DSDA Recording Config Moves Demo Files
The batch files for DSDA recordings now have commands for moving the
demo files from the 'demos' directory to the source port directory, for
the duration of the running game.

This keeps the source port directory tidy and the demos a bit more
managed in their respective directories for each game. The batch files
were extended to create the map directory if it doesn't exist, rather
than having lots of map directories, most of which would remain empty.