Generate Launchers for Demos
For each demo file for any given game, a corresponding 'demo launcher'
will be generated. The 'demo-launchers' directory will basically mimic
the 'demos' directory, but instead of an lmp file, there will be a bat
file.

The `-playdemo` argument can be a direct full path to a file in a
different directory, so it points directly to the lmp file.

This is a fairly easy system to use, because any time a demo is
downloaded, it just needs to be placed in the right folder and a batch
file can be generated automatically. Whenever you want to play it again,
just launch the batch file again. No moving files around and the source
port directory will be kept clean of zillions of different demo files.