Save Game Configuration
To overcome the limitation of only having 8 save slots, rather than have
the save games stored in the same directory as the source port, we take
advantage of the `-save` argument to keep saves stored in a separate
directory for each map. I often use multiple different saves for
practicing a map, so this is useful to prevent manually moving them
around.

The saves are organised as follows:
```
<Doom Home>\saves\<WAD Name>\<Source Port>\MAPxx

Example:
C:\Users\Chris\doom\saves\2015-08-10 -- Sunlust\dsda\MAP18
```

Code was added to create the save directories for each game and to
specify the `-save` argument on each launcher file.